Move code to break dependency
