+++ programs/keeper/CMakeLists.txt
@@ -13,7 +13,6 @@ set (CLICKHOUSE_KEEPER_LINK
         clickhouse-keeper-converter-lib
         clickhouse-keeper-client-lib
         clickhouse_functions
-        clickhouse_storages_system
         dbms
 )
