+++ src/Processors/QueryPlan/Optimizations/applyOrder.cpp
@@ -9,8 +9,7 @@
 #include <Processors/QueryPlan/UnionStep.h>
 #include <Processors/QueryPlan/ReadFromMergeTree.h>
 #include <Processors/QueryPlan/SortingStep.h>
-
-#include <Storages/System/StorageSystemMetricLogView.h>
+#include <Processors/QueryPlan/CustomMetricLogViewStep.h>
 
 #include <Functions/IFunction.h>
 
@@ -43,7 +42,7 @@ SortingProperty applyOrder(QueryPlan::Node * parent, SortingProperty * propertie
     if (const auto * read_from_merge_tree = typeid_cast<ReadFromMergeTree *>(parent->step.get()))
         return {read_from_merge_tree->getSortDescription(), SortingProperty::SortScope::Stream};
 
-    if (const auto * custom_metric_log_step = typeid_cast<CustomMetricLogStep *>(parent->step.get()))
+    if (const auto * custom_metric_log_step = typeid_cast<CustomMetricLogViewStep *>(parent->step.get()))
         return {custom_metric_log_step->getSortDescription(), SortingProperty::SortScope::Global};
 
     if (const auto * aggregating_step = typeid_cast<AggregatingStep *>(parent->step.get()))
