+++ src/Processors/QueryPlan/Optimizations/removeRedundantSorting.cpp
@@ -16,8 +16,8 @@
 #include <Processors/QueryPlan/SortingStep.h>
 #include <Processors/QueryPlan/UnionStep.h>
 #include <Processors/QueryPlan/WindowStep.h>
+#include <Processors/QueryPlan/CustomMetricLogViewStep.h>
 
-#include <Storages/System/StorageSystemMetricLogView.h>
 
 #include <Common/logger_useful.h>
 #include <Common/typeid_cast.h>
@@ -67,7 +67,7 @@ public:
             || typeid_cast<FillingStep *>(current_step) /// (3) if ORDER BY is with FILL WITH, it is non-removable
             || typeid_cast<SortingStep *>(current_step) /// (4) ORDER BY will change order of previous sorting
             || typeid_cast<AggregatingStep *>(current_step) /// (5) aggregation change order
-            || typeid_cast<CustomMetricLogStep *>(current_step))
+            || typeid_cast<CustomMetricLogViewStep *>(current_step))
         {
             logStep("nodes_affect_order/push", current_node);
             nodes_affect_order.push_back(current_node);
@@ -155,7 +155,7 @@ private:
             || typeid_cast<FillingStep *>(step_affect_order))
             return false;
 
-        if (typeid_cast<CustomMetricLogStep *>(step_affect_order))
+        if (typeid_cast<CustomMetricLogViewStep *>(step_affect_order))
             return false;
 
         /// (1) aggregation
